+++ src/Services/Ordering/Ordering.API/Infrastructure/Filters/GlobalExceptionFilter.cs
@@ -37,7 +37,7 @@ namespace Ordering.Infrastructure.Filters
                 problemDetails.Errors.Add("DomainValidations", new string[] { context.Exception.Message.ToString() });
 
                 context.Result = new BadRequestObjectResult(problemDetails);
-                context.HttpContext.Response.StatusCode = (int)StatusCodes.Status400BadRequest;
+                context.HttpContext.Response.StatusCode = StatusCodes.Status400BadRequest;
 
             }
             else
@@ -49,7 +49,7 @@ namespace Ordering.Infrastructure.Filters
                 };
 
                 context.Result = new BadRequestObjectResult(errorDetails);
-                context.HttpContext.Response.StatusCode = (int)StatusCodes.Status400BadRequest;
+                context.HttpContext.Response.StatusCode = StatusCodes.Status400BadRequest;
             }
 
             context.ExceptionHandled = true;
